refactor TemporalBaseColumn find_and_replace
